Move CompilerInstaneWithContext to its own header and implementation files.
